feat: 단순 숫자 입력시 출력 성공
수는 BigInteger로 관리하여 Service에서 변환 및 검사하여 결과 출력

+++ src/main/java/calculator/view/CalculatorView.java
@@ -1,6 +1,7 @@
 package calculator.view;
 
 import camp.nextstep.edu.missionutils.Console;
+import java.math.BigInteger;
 
 public class CalculatorView {
     //TODO: 사용자로부터 입력을 받고, 계산한 뒤의 출력을 담당하는 클래스
@@ -13,7 +14,7 @@ public class CalculatorView {
         return Console.readLine();
     }
 
-    public void printResultMsg(int result) {
+    public void printResultMsg(BigInteger result) {
         System.out.printf("결과 : %d", result);
     }
 }
